requirements.txt
nextcord
python-dotenv
play2048